requirements.txt
requests
pandas
python-dotenv
flatten-json
pyarrow
ipykernel
cron-descriptor